+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## next
 
+* Support LLVM 22:
+  * `DICompileUnit'` now has an additional `dicuSourceLanguageVersion :: Word64`
+    field.
+
 * Corrected the printing of non-normal single-precision floating point
   constants.